Updated listingData object to create image array
When JSON data is added to listingData object, list of image names
is split and created as and array in the .image property.
Images for each MLS listing are in a directory with the MLS number in
img/mls/####### to make file management easier as more properties are
listed

Added the content for a new property and display sale status on the
listing item grid.